changes for interactive list
