+++ src/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">    
     <title>To Do List</title>
 </head>
 <body>
@@ -19,6 +19,6 @@
                 <button type="button" class="clearAllBtn" id="clearAllBtn" > Clear all completed </button>
             </div>
         </div>
-    </section>    
+    </section>     
 </body>
 </html>
